Apply task group sorting based on webserver config in grid structure response (#49418)
* Apply topological sort on task group grid

* Fix task_group sorting by utilizing get_task_group_children_getter to sort based on webserver config

* Adjust test expectation to sort children topologically
